# backend/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Within the scope of the Machinaide project, an API was developed using the Python language and Flask framework in line with the needs. This page contains information on how to use the API developed within the Machinaide project. This page explains how to send a request to the developed methods, which parameters will be sent, and what kind of results will be obtained.",
        "version": "1.0",
        "title": "Machinaide REST API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "suat@erstesoftware.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "vmi515134.contaboserver.net",
    "basePath": "/api/v1.0",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/factory/getFactories": {
            "get": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method contains a list of existing factories and some information about each factory.",
                "description": "This method contains a list of existing factories and some information about each factory.",
                "operationId": "getFactories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/getProductionLines": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method contains a list of existing production lines and some information about each production line.",
                "description": "This method contains a list of existing production lines and some information about each production line.",
                "operationId": "getProductionLines",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "ID of the desired factory.",
                        "required": true,
                        "properties": {
                            "factoryId": {
                                "type": "string",
                                "example": "Ermetal"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/getMachines": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method returns a list of machines belonging to a factory and some information about each machine.",
                "description": "This method returns a list of machines belonging to a factory and some information about each machine.",
                "operationId": "getMachines",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "ID of the desired factory and production line",
                        "required": true,
                        "properties": {
                            "factoryId": {
                                "type": "string",
                                "example": "Ermetal"
                            },
                            "plId": {
                                "type": "string",
                                "example": "1600T_Press_Line"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/getComponents": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method returns a list of components for a machine and some information about each component.",
                "description": "This method returns a list of components for a machine and some information about each component.",
                "operationId": "getComponents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "properties": {
                            "factoryId": {
                                "type": "string",
                                "example": "Ermetal",
                                "description": "ID of the desired factory."
                            },
                            "plId": {
                                "type": "string",
                                "example": "1600T_Press_Line",
                                "description": "ID of the desired production line"
                            },
                            "machineId": {
                                "type": "string",
                                "example": "Press030",
                                "description": "ID of the desired machine."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/getSensors": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method returns a list of sensors for a component and some information about each sensor.",
                "description": "This method returns a list of sensors for a component and some information about each sensor.",
                "operationId": "getSensors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "properties": {
                            "factoryId": {
                                "type": "string",
                                "example": "Ermetal",
                                "description": "ID of the desired factory."
                            },
                            "plId": {
                                "type": "string",
                                "example": "1600T_Press_Line",
                                "description": "ID of the desired production line"
                            },
                            "machineId": {
                                "type": "string",
                                "example": "Press031",
                                "description": "ID of the desired machine."
                            },
                            "componentId": {
                                "type": "string",
                                "example": "dengeleme",
                                "description": "ID of the desired component."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/getFields": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method returns a list of fields for a sensor and some information about each field.",
                "description": "This method returns a list of fields for a component and some information about each field.",
                "operationId": "getFields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "properties": {
                            "factoryId": {
                                "type": "string",
                                "example": "Ermetal",
                                "description": "ID of the desired factory."
                            },
                            "plId": {
                                "type": "string",
                                "example": "1600T_Press_Line",
                                "description": "ID of the desired production line"
                            },
                            "machineId": {
                                "type": "string",
                                "example": "Press031",
                                "description": "ID of the desired machine."
                            },
                            "componentId": {
                                "type": "string",
                                "example": "dengeleme",
                                "description": "ID of the desired component."
                            },
                            "sensorId": {
                                "type": "string",
                                "example": "presAnaHavaAkis",
                                "description": "ID of the desired sensor."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/factory/addMachineAction": {
            "post": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method is used to add a new machine action record.",
                "description": "This method is used to add a new machine action record.",
                "operationId": "addMachineAction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MachineAction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/factory/getMaterials": {
            "get": {
                "tags": [
                    "Factory"
                ],
                "summary": "This method contains a list of existing materials and some information about each material.",
                "description": "This method contains a list of existing materials and some information about each material.",
                "operationId": "getMaterials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "Token": {
            "type": "string",
            "format": "string",
            "example": ""
        },
...
    },
...
}